Add E. Binary Inversions
This problem is a good example of how to use prefix sums.

The solution is to flip the first 0 and the last 1.

The first 0 is flipped to 1 because it will increase the number of inversions
by the number of 1s before it.

The last 1 is flipped to 0 because it will decrease the number of inversions
by the number of 1s after it.

The number of inversions is the number of 1s before the first 0 plus the
number of 1s after the last 1.