Fixing the testcase
The testcase was not being read properly.

The testcase was being read as a string, but it was supposed to be read as a list.

The testcase was being read as a string, but it was supposed to be read as a list.

The testcase was being read as a string, but it was supposed to be read as a list.