Fix typo in A_Problem-_1.py
The problem statement says that the answer is the number of ways to choose
a subset of the 1s such that the sum of the indices of the chosen 1s is
divisible by 2^k.

The code was computing the number of ways to choose a subset of the 1s
such that the sum of the indices of the chosen 1s is divisible by 2^k
times the number of 1s.